backend/requirements.txt
blinker==1.9.0
click==8.1.7
colorama==0.4.6
dnspython==1.16.0
et_xmlfile==2.0.0
Flask==3.1.0
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.3
openpyxl==3.1.5
pandas==2.2.3
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
tzdata==2024.2
Werkzeug==3.1.3
